Generic collections in dasbhoard-displayer-core

+++ modules/dashboard-displayers/dashboard-displayer-core/src/main/java/org/jboss/dashboard/displayer/table/Table.java
@@ -136,7 +136,7 @@ public class Table {
     }
 
     public int getNumberOfPages() {
-        return (int) ((getRowCount() - 1) / getMaxRowsPerPage() + 1);
+        return ((getRowCount() - 1) / getMaxRowsPerPage() + 1);
     }
 
     public int getCurrentPage() {
